refactor: extract server-skills.js from server.js (PR-17)
Move skills proxy cache and routes into dedicated server-skills
module: httpGet, fetchSkillsPage, fetchSkillDetail, skillsCache,
permission gate, leaderboard list, search, and detail endpoints.

server.js delegates skills routes via skills.handleRequest().

server.js: 2428 -> 2199 lines. server-skills.js: 280 lines.
